pyproject.toml
[tool.poetry]
name = "kiu"
version = "0.1.0"
description = "Kiu airlines solutions"
authors = ["Cristian Barreto <cristianbarretof@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.2"
djangorestframework = "^3.15.2"
coverage = "^7.6.4"
psycopg2-binary = "^2.9.10"
requests = "^2.32.3"

[tool.poetry.group.dev.dependencies]
ipython = "^8.29.0"
ipdb = "^0.13.13"
django-extensions = "^3.2.3"
nplusone = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
